Create 256 BIT AES encryption.
Here is the first code you will be running to encrypt any string or text. You will make a open end passphrase to start the encryption. This will be in other words your public key of your choice and you will be asked to write a message. In writing your message you will receive the encrypted message and the decryption key which was written from base64.